--- src/js/electron/tron/index.ts
@@ -1,9 +1,0 @@
-import session from "./session"
-import window from "./window"
-
-export type Session = ReturnType<typeof session>
-
-export default {
-  window,
-  session,
-}
