+++ src/js/electron/windows/about-window.ts
@@ -1,4 +1,4 @@
-import {WindowName} from "../tron/window-manager"
+import {WindowName} from "../windows/window-manager"
 import {ZuiWindow} from "./zui-window"
 
 export class AboutWindow extends ZuiWindow {
