+++ src/js/electron/windows/dimens.test.ts
