+++ src/js/electron/windows/dimens.ts
