+++ src/js/electron/windows/window-manager.test.ts
