+++ src/js/electron/windows/window-manager.ts
@@ -2,8 +2,8 @@ import {BrowserWindow, screen} from "electron"
 import log from "electron-log"
 import {last} from "lodash"
 import {NewTabSearchParams} from "../ipc/windows/messages"
-import {stack} from "../window/dimens"
-import {SessionState} from "./session-state"
+import {stack} from "../windows/dimens"
+import {SessionState} from "../session-state"
 import {State} from "src/js/state/types"
 import {ZuiWindow} from "../windows/zui-window"
 import {HiddenWindow} from "../windows/hidden-window"
