+++ src/js/electron/windows/zui-window.ts
@@ -2,8 +2,8 @@ import {enable} from "@electron/remote/main"
 import {nanoid} from "@reduxjs/toolkit"
 import {BrowserWindow, BrowserWindowConstructorOptions} from "electron"
 import {State} from "src/js/state/types"
-import {WindowName} from "../tron/window-manager"
-import {Dimens, getWindowDimens, pickDimens} from "../window/dimens"
+import {WindowName} from "../windows/window-manager"
+import {Dimens, getWindowDimens, pickDimens} from "../windows/dimens"
 import {getDisplays} from "./get-displays"
 import {SerializedWindow, WindowProps} from "./types"
